refactor: rearrange test classes
